Issue #51: Seperate Favoriting logic out of Topic viewing
